refactor: modularize codebase and condense PR template - Extract search service logic into focused modules (filtering.go, multi_search.go) - Split API handlers into separate files by functionality - Decompose engine into focused modules - Condense GitHub PR template from 162 to 70 lines (57% reduction) - Maintain all functionality and test coverage - Improve code organization and maintainability
