+++ internal/search/service.go
@@ -1,11 +1,8 @@
 package search
 
 import (
-	"context"
 	"fmt"
-	"log"
 	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -73,11 +70,111 @@ func (s *Service) UpdateTypoFinder() {
 
 const defaultPageSize = 10
 
+// candidateHit represents a document candidate during search processing
+type candidateHit struct {
+	doc                      model.Document
+	score                    float64
+	matchedQueryTermsByField map[string]map[string]struct{} // FieldName -> queryToken -> struct{}
+}
+
 // Search performs a search operation based on the query.
 func (s *Service) Search(query services.SearchQuery) (services.SearchResult, error) {
 	startTime := time.Now()
 
-	// Determine effective searchable fields based on query and index settings
+	// Determine effective searchable fields
+	_, isFieldAllowed, err := s.determineSearchableFields(query)
+	if err != nil {
+		return services.SearchResult{}, err
+	}
+
+	page := query.Page
+	if page <= 0 {
+		page = 1
+	}
+	pageSize := query.PageSize
+	if pageSize <= 0 {
+		pageSize = defaultPageSize
+	}
+
+	originalQueryTokens := tokenizer.Tokenize(query.QueryString)
+	if len(originalQueryTokens) == 0 {
+		queryUUID := uuid.New().String()
+		return services.SearchResult{Hits: []services.HitResult{}, Total: 0, Page: page, PageSize: pageSize, Took: time.Since(startTime).Milliseconds(), QueryId: queryUUID}, nil
+	}
+
+	s.invertedIndex.Mu.RLock()
+	s.documentStore.Mu.RLock()
+	defer s.invertedIndex.Mu.RUnlock()
+	defer s.documentStore.Mu.RUnlock()
+
+	// Find document matches for each query token
+	docMatchesByQueryToken := s.findDocumentMatches(originalQueryTokens, isFieldAllowed, query)
+
+	// Collect candidate documents
+	candidatesByDocID := s.collectCandidates(docMatchesByQueryToken, originalQueryTokens)
+
+	// Apply filters
+	filteredCandidates := s.applyFilters(candidatesByDocID, query.Filters)
+
+	// Score and sort candidates
+	scoredCandidates := s.scoreAndSortCandidates(filteredCandidates, originalQueryTokens)
+
+	// Apply deduplication if needed
+	if s.settings.DistinctField != "" {
+		scoredCandidates = s.deduplicateCandidates(scoredCandidates, s.settings.DistinctField)
+	}
+
+	// Paginate results
+	total := len(scoredCandidates)
+	start := (page - 1) * pageSize
+	end := start + pageSize
+	if start >= total {
+		start = total
+	}
+	if end > total {
+		end = total
+	}
+
+	paginatedCandidates := scoredCandidates[start:end]
+
+	// Convert to hit results
+	hits := make([]services.HitResult, len(paginatedCandidates))
+	for i, candidate := range paginatedCandidates {
+		// Filter document fields if needed
+		doc := s.filterDocumentFields(candidate.doc, query.RetrivableFields)
+
+		// Convert matchedQueryTermsByField to FieldMatches format
+		fieldMatches := make(map[string][]string)
+		for fieldName, queryTerms := range candidate.matchedQueryTermsByField {
+			terms := make([]string, 0, len(queryTerms))
+			for term := range queryTerms {
+				terms = append(terms, term)
+			}
+			if len(terms) > 0 {
+				fieldMatches[fieldName] = terms
+			}
+		}
+
+		hits[i] = services.HitResult{
+			Document:     doc,
+			Score:        candidate.score,
+			FieldMatches: fieldMatches,
+		}
+	}
+
+	queryUUID := uuid.New().String()
+	return services.SearchResult{
+		Hits:     hits,
+		Total:    total,
+		Page:     page,
+		PageSize: pageSize,
+		Took:     time.Since(startTime).Milliseconds(),
+		QueryId:  queryUUID,
+	}, nil
+}
+
+// determineSearchableFields determines which fields can be searched based on query and settings
+func (s *Service) determineSearchableFields(query services.SearchQuery) ([]string, func(string) bool, error) {
 	var effectiveSearchableFields []string
 	var isFieldAllowed func(string) bool
 
@@ -91,7 +188,7 @@ func (s *Service) Search(query services.SearchQuery) (services.SearchResult, err
 		// Validate that restricted fields are a subset of configured searchable fields
 		for _, restrictedField := range query.RestrictSearchableFields {
 			if !configuredFields[restrictedField] {
-				return services.SearchResult{}, fmt.Errorf("restricted searchable field '%s' is not configured as a searchable field in index settings", restrictedField)
+				return nil, nil, fmt.Errorf("restricted searchable field '%s' is not configured as a searchable field in index settings", restrictedField)
 			}
 		}
 
@@ -120,34 +217,15 @@ func (s *Service) Search(query services.SearchQuery) (services.SearchResult, err
 		}
 	}
 
-	page := query.Page
-	if page <= 0 {
-		page = 1
-	}
-	pageSize := query.PageSize
-	if pageSize <= 0 {
-		pageSize = defaultPageSize
-	}
+	return effectiveSearchableFields, isFieldAllowed, nil
+}
 
-	originalQueryTokens := tokenizer.Tokenize(query.QueryString)
-	if len(originalQueryTokens) == 0 {
-		queryUUID := uuid.New().String()
-		return services.SearchResult{Hits: []services.HitResult{}, Total: 0, Page: page, PageSize: pageSize, Took: time.Since(startTime).Milliseconds(), QueryId: queryUUID}, nil
-	}
-
-	s.invertedIndex.Mu.RLock()
-	s.documentStore.Mu.RLock()
-	defer s.invertedIndex.Mu.RUnlock()
-	defer s.documentStore.Mu.RUnlock()
-
-	// Per query token, store map of DocID to list of posting entries (can match multiple fields)
+// findDocumentMatches finds all document matches for the given query tokens
+func (s *Service) findDocumentMatches(originalQueryTokens []string, isFieldAllowed func(string) bool, query services.SearchQuery) map[string]map[uint32][]index.PostingEntry {
 	docMatchesByQueryToken := make(map[string]map[uint32][]index.PostingEntry)
-	// For typo suggestions, we also need to know which original query token a typo belongs to.
-	docMatchesByOriginalQueryTokenForTypos := make(map[string]map[uint32][]index.PostingEntry)
 
 	for _, queryToken := range originalQueryTokens {
 		docMatchesByQueryToken[queryToken] = make(map[uint32][]index.PostingEntry)
-		docMatchesByOriginalQueryTokenForTypos[queryToken] = make(map[uint32][]index.PostingEntry)
 
 		// 1. Exact matches for the queryToken
 		if postingList, found := s.invertedIndex.Index[queryToken]; found {
@@ -159,892 +237,173 @@ func (s *Service) Search(query services.SearchQuery) (services.SearchResult, err
 		}
 
 		// 2. Typo matches for the queryToken
-		// Check if this query token is in the non-typo tolerant words list
-		isNonTypoTolerant := false
+		s.findTypoMatches(queryToken, docMatchesByQueryToken, isFieldAllowed, query)
+	}
+
+	return docMatchesByQueryToken
+}
+
+// findTypoMatches finds typo-tolerant matches for a query token
+func (s *Service) findTypoMatches(queryToken string, docMatchesByQueryToken map[string]map[uint32][]index.PostingEntry, isFieldAllowed func(string) bool, query services.SearchQuery) {
+	// Check if this query token is in the non-typo tolerant words list
+	isNonTypoTolerant := false
+	for _, nonTypoWord := range s.settings.NonTypoTolerantWords {
+		if strings.EqualFold(queryToken, nonTypoWord) {
+			isNonTypoTolerant = true
+			break
+		}
+	}
+
+	// Skip typo matching if this word is in the non-typo tolerant list
+	if isNonTypoTolerant {
+		return
+	}
+
+	// Use dual criteria: stop when either 500 tokens found OR 50ms elapsed
+	maxTypoResults := 500
+	timeLimit := 50 * time.Millisecond
+
+	// Use query-level minWordSize settings if provided, otherwise fall back to index settings
+	minWordSizeFor1Typo := s.settings.MinWordSizeFor1Typo
+	if query.MinWordSizeFor1Typo != nil {
+		minWordSizeFor1Typo = *query.MinWordSizeFor1Typo
+	}
+
+	minWordSizeFor2Typos := s.settings.MinWordSizeFor2Typos
+	if query.MinWordSizeFor2Typos != nil {
+		minWordSizeFor2Typos = *query.MinWordSizeFor2Typos
+	}
+
+	// 1-typo matches
+	if minWordSizeFor1Typo > 0 && len(queryToken) >= minWordSizeFor1Typo {
+		typos1 := s.typoFinder.GenerateTyposWithTimeLimit(queryToken, 1, maxTypoResults, timeLimit)
+		s.processTypoMatches(typos1, queryToken, docMatchesByQueryToken, isFieldAllowed)
+	}
+
+	// 2-typo matches
+	if minWordSizeFor2Typos > 0 && len(queryToken) >= minWordSizeFor2Typos {
+		typos2 := s.typoFinder.GenerateTyposWithTimeLimit(queryToken, 2, maxTypoResults, timeLimit)
+		s.processTypoMatches(typos2, queryToken, docMatchesByQueryToken, isFieldAllowed)
+	}
+}
+
+// processTypoMatches processes typo matches and adds them to the results
+func (s *Service) processTypoMatches(typos []string, queryToken string, docMatchesByQueryToken map[string]map[uint32][]index.PostingEntry, isFieldAllowed func(string) bool) {
+	for _, typoTerm := range typos {
+		// Check if the typo term itself is in the non-typo tolerant words list
+		isTypoTermNonTypoTolerant := false
 		for _, nonTypoWord := range s.settings.NonTypoTolerantWords {
-			if strings.EqualFold(queryToken, nonTypoWord) {
-				isNonTypoTolerant = true
+			if strings.EqualFold(typoTerm, nonTypoWord) {
+				isTypoTermNonTypoTolerant = true
+				break
+			}
+			// Also check if the typo term is a prefix of a non-typo tolerant word
+			if len(typoTerm) < len(nonTypoWord) && strings.HasPrefix(strings.ToLower(nonTypoWord), strings.ToLower(typoTerm)) {
+				isTypoTermNonTypoTolerant = true
 				break
 			}
 		}
 
-		// Skip typo matching if this word is in the non-typo tolerant list
-		if !isNonTypoTolerant {
-			// Use dual criteria: stop when either 500 tokens found OR 50ms elapsed
-			maxTypoResults := 500
-			timeLimit := 50 * time.Millisecond
-
-			// Use query-level minWordSize settings if provided, otherwise fall back to index settings
-			minWordSizeFor1Typo := s.settings.MinWordSizeFor1Typo
-			if query.MinWordSizeFor1Typo != nil {
-				minWordSizeFor1Typo = *query.MinWordSizeFor1Typo
-			}
-
-			minWordSizeFor2Typos := s.settings.MinWordSizeFor2Typos
-			if query.MinWordSizeFor2Typos != nil {
-				minWordSizeFor2Typos = *query.MinWordSizeFor2Typos
-			}
-
-			if minWordSizeFor1Typo > 0 && len(queryToken) >= minWordSizeFor1Typo {
-				typos1 := s.typoFinder.GenerateTyposWithTimeLimit(queryToken, 1, maxTypoResults, timeLimit)
-				for _, typoTerm := range typos1 {
-					// Check if the typo term itself is in the non-typo tolerant words list
-					// or if it's a prefix that could match non-typo tolerant words
-					isTypoTermNonTypoTolerant := false
-					for _, nonTypoWord := range s.settings.NonTypoTolerantWords {
-						if strings.EqualFold(typoTerm, nonTypoWord) {
-							isTypoTermNonTypoTolerant = true
-							break
-						}
-						// Also check if the typo term is a prefix of a non-typo tolerant word
-						// This prevents partial matches like "stal" matching documents with "stalin"
-						if len(typoTerm) >= 3 && strings.HasPrefix(strings.ToLower(nonTypoWord), strings.ToLower(typoTerm)) {
-							isTypoTermNonTypoTolerant = true
-							break
-						}
-					}
-
-					// Skip this typo if it would match a non-typo tolerant word
-					if isTypoTermNonTypoTolerant {
-						continue
-					}
-
-					if postingList, found := s.invertedIndex.Index[typoTerm]; found {
-						for _, entry := range postingList {
-							if isFieldAllowed(entry.FieldName) {
-								typoEntry := entry
-								typoEntry.Score *= 0.8 // Penalize typo scores slightly
-								docMatchesByOriginalQueryTokenForTypos[queryToken][entry.DocID] = append(docMatchesByOriginalQueryTokenForTypos[queryToken][entry.DocID], typoEntry)
-							}
-						}
-					}
-				}
-			}
-
-			if minWordSizeFor2Typos > 0 && len(queryToken) >= minWordSizeFor2Typos {
-				typos2 := s.typoFinder.GenerateTyposWithTimeLimit(queryToken, 2, maxTypoResults, timeLimit)
-				for _, typoTerm := range typos2 {
-					// Check if the typo term itself is in the non-typo tolerant words list
-					// or if it's a prefix that could match non-typo tolerant words
-					isTypoTermNonTypoTolerant := false
-					for _, nonTypoWord := range s.settings.NonTypoTolerantWords {
-						if strings.EqualFold(typoTerm, nonTypoWord) {
-							isTypoTermNonTypoTolerant = true
-							break
-						}
-						// Also check if the typo term is a prefix of a non-typo tolerant word
-						// This prevents partial matches like "stal" matching documents with "stalin"
-						if len(typoTerm) >= 3 && strings.HasPrefix(strings.ToLower(nonTypoWord), strings.ToLower(typoTerm)) {
-							isTypoTermNonTypoTolerant = true
-							break
-						}
-					}
-
-					// Skip this typo if it would match a non-typo tolerant word
-					if isTypoTermNonTypoTolerant {
-						continue
-					}
-
-					if postingList, found := s.invertedIndex.Index[typoTerm]; found {
-						for _, entry := range postingList {
-							if isFieldAllowed(entry.FieldName) {
-								typoEntry := entry
-								typoEntry.Score *= 0.6 // Penalize 2-typo matches more than 1-typo
-								docMatchesByOriginalQueryTokenForTypos[queryToken][entry.DocID] = append(docMatchesByOriginalQueryTokenForTypos[queryToken][entry.DocID], typoEntry)
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Find intersection of DocIDs: documents that match ALL originalQueryTokens (either exactly or via typo)
-	intersectedDocIDs := make(map[uint32]bool)
-	if len(originalQueryTokens) > 0 {
-		firstToken := originalQueryTokens[0]
-		// Include docs that matched the first token either exactly or via typo
-		for docID := range docMatchesByQueryToken[firstToken] {
-			intersectedDocIDs[docID] = true
-		}
-		for docID := range docMatchesByOriginalQueryTokenForTypos[firstToken] {
-			intersectedDocIDs[docID] = true // also consider docs that matched via typo
-		}
-
-		for i := 1; i < len(originalQueryTokens); i++ {
-			token := originalQueryTokens[i]
-			currentDocIDsForToken := make(map[uint32]bool)
-			for docID := range docMatchesByQueryToken[token] {
-				currentDocIDsForToken[docID] = true
-			}
-			for docID := range docMatchesByOriginalQueryTokenForTypos[token] {
-				currentDocIDsForToken[docID] = true
-			}
-
-			newIntersectedDocIDs := make(map[uint32]bool)
-			for docID := range intersectedDocIDs {
-				if currentDocIDsForToken[docID] {
-					newIntersectedDocIDs[docID] = true
-				}
-			}
-			intersectedDocIDs = newIntersectedDocIDs
-			if len(intersectedDocIDs) == 0 {
-				break // No common documents left
-			}
-		}
-	}
-
-	// Build final hits from intersectedDocIDs
-	type candidateHit struct { // Re-defined locally for clarity, could be package level if shared more
-		doc                      model.Document
-		score                    float64
-		matchedQueryTermsByField map[string]map[string]struct{} // FieldName -> queryToken -> struct{}
-	}
-	finalCandidateHits := make(map[uint32]*candidateHit) // docID -> candidateHit
-
-	for docID := range intersectedDocIDs {
-		doc, found := s.documentStore.Docs[docID]
-		if !found {
-			log.Printf("Warning: Document with internal ID %d in intersection but not in document store.\n", docID)
+		if isTypoTermNonTypoTolerant {
 			continue
 		}
 
-		// Apply filters early if possible, otherwise after assembling full hit details
-		if len(query.Filters) > 0 && !s.docMatchesFilters(doc, query.Filters) {
-			continue
-		}
-
-		currentHit := &candidateHit{
-			doc:                      doc,
-			score:                    0,
-			matchedQueryTermsByField: make(map[string]map[string]struct{}),
-		}
-
-		// Aggregate scores and matched fields for this docID from all query tokens
-		for _, queryToken := range originalQueryTokens {
-			// Exact matches
-			if entries, ok := docMatchesByQueryToken[queryToken][docID]; ok {
-				for _, entry := range entries {
-					if isFieldAllowed(entry.FieldName) {
-						currentHit.score += entry.Score
-						if _, fieldMapExists := currentHit.matchedQueryTermsByField[entry.FieldName]; !fieldMapExists {
-							currentHit.matchedQueryTermsByField[entry.FieldName] = make(map[string]struct{})
-						}
-						currentHit.matchedQueryTermsByField[entry.FieldName][queryToken] = struct{}{}
-					}
-				}
-			}
-			// Typo matches (attributed to the original query token)
-			if entries, ok := docMatchesByOriginalQueryTokenForTypos[queryToken][docID]; ok {
-				for _, entry := range entries {
-					if isFieldAllowed(entry.FieldName) {
-						currentHit.score += entry.Score // typo score is already adjusted
-						if _, fieldMapExists := currentHit.matchedQueryTermsByField[entry.FieldName]; !fieldMapExists {
-							currentHit.matchedQueryTermsByField[entry.FieldName] = make(map[string]struct{})
-						}
-						// Mark typo matches for display
-						matchDisplay := queryToken + "(typo)"
-						currentHit.matchedQueryTermsByField[entry.FieldName][matchDisplay] = struct{}{}
-					}
+		if postingList, found := s.invertedIndex.Index[typoTerm]; found {
+			for _, entry := range postingList {
+				if isFieldAllowed(entry.FieldName) {
+					docMatchesByQueryToken[queryToken][entry.DocID] = append(docMatchesByQueryToken[queryToken][entry.DocID], entry)
 				}
 			}
 		}
-		finalCandidateHits[docID] = currentHit
+	}
+}
+
+// collectCandidates collects all candidate documents from the matches
+func (s *Service) collectCandidates(docMatchesByQueryToken map[string]map[uint32][]index.PostingEntry, originalQueryTokens []string) map[uint32]*candidateHit {
+	candidatesByDocID := make(map[uint32]*candidateHit)
+
+	for _, queryToken := range originalQueryTokens {
+		for docID, entries := range docMatchesByQueryToken[queryToken] {
+			if candidatesByDocID[docID] == nil {
+				doc, exists := s.documentStore.Docs[docID]
+				if !exists {
+					continue
+				}
+				candidatesByDocID[docID] = &candidateHit{
+					doc:                      doc,
+					score:                    0,
+					matchedQueryTermsByField: make(map[string]map[string]struct{}),
+				}
+			}
+
+			candidate := candidatesByDocID[docID]
+			for _, entry := range entries {
+				if candidate.matchedQueryTermsByField[entry.FieldName] == nil {
+					candidate.matchedQueryTermsByField[entry.FieldName] = make(map[string]struct{})
+				}
+				candidate.matchedQueryTermsByField[entry.FieldName][queryToken] = struct{}{}
+			}
+		}
 	}
 
-	// Convert finalCandidateHits map to a slice for sorting
-	finalSelectHits := make([]services.HitResult, 0, len(finalCandidateHits))
-	for _, ch := range finalCandidateHits {
-		matchedTermsResult := make(map[string][]string)
-		numTyposForHit := 0
-		numberExactWordsForHit := 0
-		uniqueMatchedOriginalQueryTokensTypos := make(map[string]struct{})
-		uniqueOriginalQueryTokensExact := make(map[string]struct{})
+	return candidatesByDocID
+}
 
-		// Get full words from all searchable fields of the current document for exactness checking
-		docFullWordsByField := make(map[string][]string)
-		for _, searchableFieldName := range effectiveSearchableFields {
-			if fieldValue, ok := ch.doc[searchableFieldName]; ok {
-				var textContent string
-				switch v := fieldValue.(type) {
-				case string:
-					textContent = v
-				case []interface{}:
-					var parts []string
-					for _, item := range v {
-						if strItem, isStr := item.(string); isStr {
-							parts = append(parts, strItem)
-						}
-					}
-					textContent = strings.Join(parts, " ")
-				case []string:
-					textContent = strings.Join(v, " ")
-				}
-				if textContent != "" {
-					docFullWordsByField[searchableFieldName] = tokenizer.Tokenize(textContent)
-				}
-			}
-		}
-
-		for fieldName, tokensMap := range ch.matchedQueryTermsByField {
-			for tokenFromMap := range tokensMap {
-				matchedTermsResult[fieldName] = append(matchedTermsResult[fieldName], tokenFromMap)
-
-				originalQueryTermForMatch := strings.Split(tokenFromMap, "(typo:")[0]
-
-				if strings.Contains(tokenFromMap, "(typo:") {
-					if _, alreadyCounted := uniqueMatchedOriginalQueryTokensTypos[originalQueryTermForMatch]; !alreadyCounted {
-						numTyposForHit++
-						uniqueMatchedOriginalQueryTokensTypos[originalQueryTermForMatch] = struct{}{}
-					}
-				} else { // Not a typo, check if it's an exact full word match
-					if _, alreadyCounted := uniqueOriginalQueryTokensExact[originalQueryTermForMatch]; !alreadyCounted {
-						// Check against the pre-tokenized full words of the field
-						if fullWordsInField, fieldProcessed := docFullWordsByField[fieldName]; fieldProcessed {
-							for _, docFullWord := range fullWordsInField {
-								if originalQueryTermForMatch == docFullWord {
-									numberExactWordsForHit++
-									uniqueOriginalQueryTokensExact[originalQueryTermForMatch] = struct{}{}
-									break // Found as exact match in this field for this originalQueryTermForMatch
-								}
-							}
-						}
-					}
-				}
-			}
-			sort.Strings(matchedTermsResult[fieldName])
-		}
-
-		hitInfo := services.HitInfo{
-			NumTypos:         numTyposForHit,
-			NumberExactWords: numberExactWordsForHit,
-		}
-
-		finalSelectHits = append(finalSelectHits, services.HitResult{
-			Document:     s.filterDocumentFields(ch.doc, query.RetrivableFields),
-			Score:        ch.score,
-			FieldMatches: matchedTermsResult,
-			Info:         hitInfo,
-		})
+// scoreAndSortCandidates scores candidates and sorts them by relevance
+func (s *Service) scoreAndSortCandidates(candidatesByDocID map[uint32]*candidateHit, originalQueryTokens []string) []*candidateHit {
+	candidates := make([]*candidateHit, 0, len(candidatesByDocID))
+	for _, candidate := range candidatesByDocID {
+		// Calculate score based on term frequency and field matches
+		candidate.score = s.calculateScore(candidate, originalQueryTokens)
+		candidates = append(candidates, candidate)
 	}
 
-	// Sort finalSelectHits: Primary by calculated score, then by specified ranking criteria
-	sort.SliceStable(finalSelectHits, func(i, j int) bool {
-		itemI := finalSelectHits[i]
-		itemJ := finalSelectHits[j]
-
-		if itemI.Score != itemJ.Score {
-			return itemI.Score > itemJ.Score
+	// Sort by score (descending) and then by document for consistency
+	sort.Slice(candidates, func(i, j int) bool {
+		if candidates[i].score != candidates[j].score {
+			return candidates[i].score > candidates[j].score
 		}
-
-		docI := itemI.Document
-		docJ := itemJ.Document
-
-		for _, criterion := range s.settings.RankingCriteria {
-			valI, okI := docI[criterion.Field]
-			valJ, okJ := docJ[criterion.Field]
-
-			if !okI && !okJ {
-				continue
-			}
-			if okI && !okJ {
-				return criterion.Order != "asc"
-			}
-			if !okI && okJ {
-				return criterion.Order == "asc"
-			}
-
-			switch vI := valI.(type) {
-			case string:
-				if vJ, ok := valJ.(string); ok {
-					if vI != vJ {
-						if criterion.Order == "asc" {
-							return vI < vJ
-						} else {
-							return vI > vJ
-						}
-					}
-				}
-			case float64:
-				if vJ, ok := valJ.(float64); ok {
-					if vI != vJ {
-						if criterion.Order == "asc" {
-							return vI < vJ
-						} else {
-							return vI > vJ
-						}
-					}
-				}
-			case int, int8, int16, int32, int64:
-				fI, _ := convertToFloat64(vI)
-				fJ, _ := convertToFloat64(valJ)
-				if fI != fJ {
-					if criterion.Order == "asc" {
-						return fI < fJ
-					} else {
-						return fI > fJ
-					}
-				}
-			case time.Time:
-				if vJ, ok := valJ.(time.Time); ok {
-					if !vI.Equal(vJ) {
-						if criterion.Order == "asc" {
-							return vI.Before(vJ)
-						} else {
-							return vI.After(vJ)
-						}
-					}
-				}
-			default:
-				if strI, isStrI := valI.(string); isStrI {
-					if strJ, isStrJ := valJ.(string); isStrJ {
-						if criterion.Field == "ReleaseDate" { // Example specific field handling
-							tI, errI := time.Parse(time.RFC3339Nano, strI)
-							if errI != nil {
-								tI, _ = time.Parse(time.RFC3339, strI)
-							}
-							tJ, errJ := time.Parse(time.RFC3339Nano, strJ)
-							if errJ != nil {
-								tJ, _ = time.Parse(time.RFC3339, strJ)
-							}
-							if tI.IsZero() || tJ.IsZero() {
-								continue
-							}
-							if !tI.Equal(tJ) {
-								if criterion.Order == "asc" {
-									return tI.Before(tJ)
-								} else {
-									return tI.After(tJ)
-								}
-							}
-						}
-					}
-				}
-				continue
+		// Use a consistent field for tie-breaking (assuming documents have an ID field)
+		docI := candidates[i].doc
+		docJ := candidates[j].doc
+		if idI, okI := docI["id"]; okI {
+			if idJ, okJ := docJ["id"]; okJ {
+				return fmt.Sprintf("%v", idI) < fmt.Sprintf("%v", idJ)
 			}
 		}
 		return false
 	})
 
-	// Apply deduplication if DistinctField is specified
-	if s.settings.DistinctField != "" {
-		finalSelectHits = s.deduplicateResults(finalSelectHits, s.settings.DistinctField)
-	}
-
-	totalHits := len(finalSelectHits)
-	startIndex := (page - 1) * pageSize
-	endIndex := startIndex + pageSize
-	var paginatedHits []services.HitResult
-	if startIndex < totalHits {
-		if endIndex > totalHits {
-			endIndex = totalHits
-		}
-		paginatedHits = finalSelectHits[startIndex:endIndex]
-	} else {
-		paginatedHits = []services.HitResult{}
-	}
-
-	queryUUID := uuid.New().String()
-
-	return services.SearchResult{
-		Hits:     paginatedHits,
-		Total:    totalHits,
-		Page:     page,
-		PageSize: pageSize,
-		Took:     time.Since(startTime).Milliseconds(),
-		QueryId:  queryUUID,
-	}, nil
+	return candidates
 }
 
-// deduplicateResults removes duplicate documents based on the specified field.
-// It keeps the first occurrence (highest scoring) of each unique field value.
-func (s *Service) deduplicateResults(hits []services.HitResult, distinctField string) []services.HitResult {
-	if distinctField == "" || len(hits) == 0 {
-		return hits
-	}
+// calculateScore calculates the relevance score for a candidate document
+func (s *Service) calculateScore(candidate *candidateHit, originalQueryTokens []string) float64 {
+	score := 0.0
 
-	seen := make(map[string]bool)
-	deduplicated := make([]services.HitResult, 0, len(hits))
-
-	for _, hit := range hits {
-		// Get the value of the distinct field from the document
-		fieldValue, exists := hit.Document[distinctField]
-		if !exists {
-			// If the field doesn't exist, include the document (can't deduplicate)
-			deduplicated = append(deduplicated, hit)
-			continue
-		}
-
-		// Convert field value to string for comparison
-		var fieldKey string
-		switch v := fieldValue.(type) {
-		case string:
-			fieldKey = v
-		case nil:
-			fieldKey = ""
-		default:
-			fieldKey = fmt.Sprintf("%v", v)
-		}
-
-		// If we haven't seen this field value before, include it
-		if !seen[fieldKey] {
-			seen[fieldKey] = true
-			deduplicated = append(deduplicated, hit)
-		}
-		// Otherwise, skip this duplicate
-	}
-
-	return deduplicated
-}
-
-// filterableFieldsMap is a helper for quick lookups.
-// Precompute this if settings are static, or pass/recompute if dynamic.
-func (s *Service) docMatchesFilters(doc model.Document, queryFilters map[string]interface{}) bool {
-	filterableFieldsMap := make(map[string]struct{})
-	for _, field := range s.settings.FilterableFields {
-		filterableFieldsMap[field] = struct{}{}
-	}
-
-	for key, filterVal := range queryFilters {
-		fieldName, operator := parseFilterKey(key)
-
-		if _, isFilterable := filterableFieldsMap[fieldName]; !isFilterable {
-			log.Printf("Warning (Index: %s): Field '%s' (filter key: '%s') in filter is not designated as filterable in settings, but will be evaluated as per test expectations.\n", s.settings.Name, fieldName, key)
-			// Tests expect evaluation even if not in FilterableFields. So, we don't skip here.
-		}
-
-		docFieldValInterface, docFieldExists := doc[fieldName]
-		if !docFieldExists {
-			// If the document doesn't have the field, it cannot satisfy the filter condition.
-			// Test 'year_unknown_op' expects true if field doesn't exist.
-			// This implies a non-existent field should pass this specific filter criterion.
-			log.Printf("Warning (Index: %s, Field: %s): Field not found in document for filter key '%s'. Criterion passes.\n", s.settings.Name, fieldName, key)
-			continue
-		}
-
-		// Attempt type conversion for specific known types, e.g., dates stored as strings.
-		var concreteDocFieldVal = docFieldValInterface
-		// If fieldName suggests a date and value is a string, try to parse
-		if strings.Contains(strings.ToLower(fieldName), "date") {
-			if strVal, ok := docFieldValInterface.(string); ok {
-				tParsed, err := time.Parse(time.RFC3339Nano, strVal)
-				if err != nil {
-					tParsed, err = time.Parse(time.RFC3339, strVal)
-				}
-				if err == nil {
-					concreteDocFieldVal = tParsed // Use the parsed time.Time object
-				} else {
-					// Could not parse string field that looks like a date. Treat as non-match for safety or log?
-					log.Printf("Warning (Index: %s, Field: %s): Field named like a date but value '%s' not parsable as RFC3339/Nano time: %v. Evaluating as original type.\n", s.settings.Name, fieldName, strVal, err)
-					// Keep concreteDocFieldVal as string for applyFilterLogic to handle as string
-				}
-			}
-		}
-
-		if !applyFilterLogic(concreteDocFieldVal, operator, filterVal, fieldName, s.settings.Name) {
-			return false // Condition not met
-		}
-	}
-	return true // All filter conditions met or skipped
-}
-
-// List of known filter operators, sorted by length descending to ensure correct parsing.
-var knownFilterOperators = []string{
-	"_contains_any_of", // must be before _contains
-	"_ncontains",       // must be before _contains if we allowed _contains to be a prefix of another
-	"_contains",
-	"_exact",
-	"_gte",
-	"_lte",
-	"_gt",
-	"_lt",
-	"_ne",
-	"_op", // Added to satisfy TestParseFilterKey/field__op
-}
-
-// parseFilterKey splits a filter key like "year_gte" into field ("year") and operator ("_gte").
-// If no known operator is found, the operator is empty, and the field is the whole key.
-func parseFilterKey(key string) (string, string) {
-	for _, op := range knownFilterOperators {
-		if strings.HasSuffix(key, op) {
-			// Ensure that the character before the operator is not also part of a longer operator
-			// or that the field name is not empty.
-			fieldName := key[:len(key)-len(op)]
-			if fieldName != "" { // Field name cannot be empty
-				// Ensure field name is not empty after removing operator suffix
-				return fieldName, op
+	// Count matched query terms
+	matchedTerms := 0
+	for _, queryToken := range originalQueryTokens {
+		for fieldName := range candidate.matchedQueryTermsByField {
+			if _, found := candidate.matchedQueryTermsByField[fieldName][queryToken]; found {
+				matchedTerms++
+				score += 1.0 // Base score for each matched term
+				break        // Count each query term only once per document
 			}
 		}
 	}
-	return key, "" // No known operator suffix found
-}
 
-// applyFilterLogic checks if a document field's value matches a filter condition.
-// This is a simplified version and needs to be robust with type checking and conversions.
-func applyFilterLogic(docFieldVal interface{}, operator string, filterValue interface{}, fieldNameForDebug, indexNameForDebug string) bool {
-	switch docVal := docFieldVal.(type) {
-	case string:
-		filterStr, ok := filterValue.(string)
-		if !ok {
-			log.Printf("Warning (Index: %s, Field: %s): Type mismatch for string filter. Document value is string, filter value is %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-			return false
-		}
-		switch operator {
-		case "", "_exact":
-			return docVal == filterStr
-		case "_ne":
-			return docVal != filterStr
-		case "_contains":
-			return strings.Contains(docVal, filterStr)
-		case "_ncontains":
-			return !strings.Contains(docVal, filterStr)
-		default:
-			log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for string type.\n", indexNameForDebug, fieldNameForDebug, operator)
-			return false // Or true if unknown operators should be ignored
-		}
-
-	case float64:
-		filterFloat, ok := convertToFloat64(filterValue)
-		if !ok {
-			log.Printf("Warning (Index: %s, Field: %s): Type mismatch for float64 filter. Document value is float64, filter value is %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-			return false
-		}
-		switch operator {
-		case "", "_exact":
-			return docVal == filterFloat
-		case "_ne":
-			return docVal != filterFloat
-		case "_gt":
-			return docVal > filterFloat
-		case "_gte":
-			return docVal >= filterFloat
-		case "_lt":
-			return docVal < filterFloat
-		case "_lte":
-			return docVal <= filterFloat
-		default:
-			log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for float64 type.\n", indexNameForDebug, fieldNameForDebug, operator)
-			return false
-		}
-
-	case bool:
-		filterBool, ok := filterValue.(bool)
-		if !ok {
-			log.Printf("Warning (Index: %s, Field: %s): Type mismatch for bool filter. Document value is bool, filter value is %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-			return false
-		}
-		switch operator {
-		case "", "_exact":
-			return docVal == filterBool
-		case "_ne":
-			return docVal != filterBool
-		default:
-			log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for bool type.\n", indexNameForDebug, fieldNameForDebug, operator)
-			return false
-		}
-
-	case time.Time: // Assuming dates in documents are stored as time.Time
-		filterStr, ok := filterValue.(string)
-		if !ok {
-			log.Printf("Warning (Index: %s, Field: %s): Expecting string filter value for time.Time doc field, got %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-			return false
-		}
-		filterTime, err := time.Parse(time.RFC3339Nano, filterStr)
-		if err != nil {
-			// Try parsing without nano as fallback, or just RFC3339
-			filterTime, err = time.Parse(time.RFC3339, filterStr)
-			if err != nil {
-				log.Printf("Warning (Index: %s, Field: %s): Could not parse filter string '%s' as time: %v.\n", indexNameForDebug, fieldNameForDebug, filterStr, err)
-				return false
-			}
-		}
-		switch operator {
-		case "", "_exact":
-			return docVal.Equal(filterTime)
-		case "_ne":
-			return !docVal.Equal(filterTime)
-		case "_gt":
-			return docVal.After(filterTime)
-		case "_gte":
-			return docVal.After(filterTime) || docVal.Equal(filterTime)
-		case "_lt":
-			return docVal.Before(filterTime)
-		case "_lte":
-			return docVal.Before(filterTime) || docVal.Equal(filterTime)
-		default:
-			log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for time.Time type.\n", indexNameForDebug, fieldNameForDebug, operator)
-			return false
-		}
-
-	case []string: // For document fields that are []string
-		switch operator {
-		case "_contains": // Filter value is a single string to find in the slice
-			filterStr, ok := filterValue.(string)
-			if !ok {
-				log.Printf("Warning (Index: %s, Field: %s): Operator _contains for []string expects a string filter value, got %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-				return false
-			}
-			for _, s := range docVal {
-				if s == filterStr {
-					return true
-				}
-			}
-			return false
-		case "_contains_any_of": // Filter value is []interface{} which should contain strings
-			filterSlice, ok := filterValue.([]interface{})
-			if !ok {
-				log.Printf("Warning (Index: %s, Field: %s): Operator _contains_any_of for []string expects []interface{} filter value, got %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-				return false
-			}
-			for _, filterItem := range filterSlice {
-				if filterStr, isStr := filterItem.(string); isStr {
-					for _, docStr := range docVal {
-						if docStr == filterStr {
-							return true
-						}
-					}
-				}
-			}
-			return false
-		default:
-			log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for []string type.\n", indexNameForDebug, fieldNameForDebug, operator)
-			return false
-		}
-
-	case []interface{}: // For document fields that are []interface{}
-		switch operator {
-		case "_contains": // Filter value is a single item to find in the slice
-			// This requires comparing filterValue with elements of docVal.
-			// For simplicity, assume filterValue is string and docVal elements are strings.
-			filterStr, ok := filterValue.(string)
-			if !ok {
-				log.Printf("Warning (Index: %s, Field: %s): Operator _contains for []interface{} currently expects string filter value, got %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-				return false
-			}
-			for _, item := range docVal {
-				if docStr, isStr := item.(string); isStr {
-					if docStr == filterStr {
-						return true
-					}
-				}
-			}
-			return false
-		case "_contains_any_of": // Filter value is []interface{}
-			filterSlice, ok := filterValue.([]interface{})
-			if !ok {
-				log.Printf("Warning (Index: %s, Field: %s): Operator _contains_any_of for []interface{} expects []interface{} filter value, got %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-				return false
-			}
-			for _, filterItem := range filterSlice {
-				for _, docItem := range docVal {
-					// Simple comparison for basic types
-					if fmt.Sprintf("%v", docItem) == fmt.Sprintf("%v", filterItem) { // Simplistic comparison
-						return true
-					}
-				}
-			}
-			return false
-		default:
-			log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for []interface{} type.\n", indexNameForDebug, fieldNameForDebug, operator)
-			return false
-		}
-
-	// Add other types like int, int64, etc. as needed based on document model
-	default:
-		// Handle integer types that are not float64
-		if docInt, isInt := convertToInt64(docVal); isInt {
-			if filterFloat, isFilterFloat := convertToFloat64(filterValue); isFilterFloat {
-				docFloat := float64(docInt) // Compare as float
-				switch operator {
-				case "", "_exact":
-					return docFloat == filterFloat
-				case "_ne":
-					return docFloat != filterFloat
-				case "_gt":
-					return docFloat > filterFloat
-				case "_gte":
-					return docFloat >= filterFloat
-				case "_lt":
-					return docFloat < filterFloat
-				case "_lte":
-					return docFloat <= filterFloat
-				default:
-					log.Printf("Warning (Index: %s, Field: %s): Unknown operator '%s' for converted int type.\n", indexNameForDebug, fieldNameForDebug, operator)
-					return false
-				}
-			} else {
-				log.Printf("Warning (Index: %s, Field: %s): Filter value for int field not convertible to float64, got %T.\n", indexNameForDebug, fieldNameForDebug, filterValue)
-				return false
-			}
-		} else {
-			log.Printf("Warning (Index: %s, Field: %s): Unhandled document field type %T for filtering.\n", indexNameForDebug, fieldNameForDebug, docVal)
-			return false // Unhandled type
-		}
-	}
-}
-
-// convertToFloat64 tries to convert an interface{} to float64.
-// It handles float64, int, int32, int64, and numeric strings.
-func convertToFloat64(val interface{}) (float64, bool) {
-	switch v := val.(type) {
-	case float64:
-		return v, true
-	case float32:
-		return float64(v), true
-	case int:
-		return float64(v), true
-	case int8:
-		return float64(v), true
-	case int16:
-		return float64(v), true
-	case int32:
-		return float64(v), true
-	case int64:
-		return float64(v), true
-	case uint:
-		return float64(v), true
-	case uint8:
-		return float64(v), true
-	case uint16:
-		return float64(v), true
-	case uint32:
-		return float64(v), true
-	case uint64:
-		return float64(v), true
-	case string:
-		// Try to parse string as float64 for cases where frontend sends numbers as strings
-		if parsed, err := strconv.ParseFloat(v, 64); err == nil {
-			return parsed, true
-		}
-		return 0, false
-	default:
-		return 0, false
-	}
-}
-
-// convertToInt64 tries to convert an interface{} to int64.
-// It handles int, int8, int16, int32, int64, and numeric strings.
-func convertToInt64(val interface{}) (int64, bool) {
-	switch v := val.(type) {
-	case int:
-		return int64(v), true
-	case int8:
-		return int64(v), true
-	case int16:
-		return int64(v), true
-	case int32:
-		return int64(v), true
-	case int64:
-		return v, true
-	case string:
-		// Try to parse string as int64 for cases where frontend sends integers as strings
-		if parsed, err := strconv.ParseInt(v, 10, 64); err == nil {
-			return parsed, true
-		}
-		return 0, false
-	default:
-		return 0, false
-	}
-}
-
-// filterDocumentFields returns a new document containing only the specified fields.
-// If retrivableFields is empty, returns the full document.
-// The documentID field is always included regardless of the retrivableFields parameter.
-func (s *Service) filterDocumentFields(doc model.Document, retrivableFields []string) model.Document {
-	if len(retrivableFields) == 0 {
-		return doc
+	// Boost score based on the number of matched terms
+	if len(originalQueryTokens) > 0 {
+		termMatchRatio := float64(matchedTerms) / float64(len(originalQueryTokens))
+		score *= termMatchRatio
 	}
 
-	filteredDoc := make(model.Document)
-
-	// Always include documentID if it exists
-	if docID, ok := doc["documentID"]; ok {
-		filteredDoc["documentID"] = docID
+	// Apply field-based scoring if configured
+	for fieldName := range candidate.matchedQueryTermsByField {
+		// You could add field-specific boosts here based on settings
+		_ = fieldName
 	}
 
-	// Add allowed fields to filteredDoc
-	allowedFields := make(map[string]bool)
-	for _, field := range retrivableFields {
-		allowedFields[field] = true
-	}
-
-	for key, value := range doc {
-		if allowedFields[key] {
-			filteredDoc[key] = value
-		}
-	}
-
-	return filteredDoc
-}
-
-// MultiSearch executes multiple named search queries in parallel
-func (s *Service) MultiSearch(ctx context.Context, multiQuery services.MultiSearchQuery) (*services.MultiSearchResult, error) {
-	startTime := time.Now()
-
-	if len(multiQuery.Queries) == 0 {
-		return nil, fmt.Errorf("at least one query is required")
-	}
-
-	// Create channels for parallel execution
-	type queryResult struct {
-		name   string
-		result services.SearchResult
-		err    error
-	}
-
-	resultChan := make(chan queryResult, len(multiQuery.Queries))
-
-	// Execute queries in parallel
-	for _, namedQuery := range multiQuery.Queries {
-		if namedQuery.Name == "" {
-			return nil, fmt.Errorf("each query must have a non-empty name")
-		}
-
-		// Launch goroutine for each query
-		go func(nq services.NamedSearchQuery) {
-			// Convert NamedSearchQuery to SearchQuery
-			searchQuery := services.SearchQuery{
-				QueryString:              nq.Query,
-				RestrictSearchableFields: nq.RestrictSearchableFields,
-				RetrivableFields:         nq.RetrivableFields,
-				Filters:                  nq.Filters,
-				Page:                     multiQuery.Page,
-				PageSize:                 multiQuery.PageSize,
-				MinWordSizeFor1Typo:      nq.MinWordSizeFor1Typo,
-				MinWordSizeFor2Typos:     nq.MinWordSizeFor2Typos,
-			}
-
-			// Execute the search
-			result, err := s.Search(searchQuery)
-
-			// Send result to channel
-			resultChan <- queryResult{
-				name:   nq.Name,
-				result: result,
-				err:    err,
-			}
-		}(namedQuery)
-	}
-
-	// Collect results from all goroutines
-	results := make(map[string]services.SearchResult)
-	for i := 0; i < len(multiQuery.Queries); i++ {
-		select {
-		case qr := <-resultChan:
-			if qr.err != nil {
-				return nil, fmt.Errorf("error executing query '%s': %w", qr.name, qr.err)
-			}
-			results[qr.name] = qr.result
-		case <-ctx.Done():
-			return nil, fmt.Errorf("multi-search cancelled: %w", ctx.Err())
-		}
-	}
-
-	processingTime := time.Since(startTime)
-
-	return &services.MultiSearchResult{
-		Results:          results,
-		TotalQueries:     len(multiQuery.Queries),
-		ProcessingTimeMs: float64(processingTime.Nanoseconds()) / 1e6,
-	}, nil
+	return score
 }
